+++ scripts/layering/package-boundaries.test.ts
@@ -434,6 +434,7 @@ test('the real tree parses, declares, and passes R11', () => {
     '@agent-device/capture-kit/snapshot-quality-verdict',
     '@agent-device/capture-kit/snapshot-rect-projection',
     '@agent-device/capture-kit/snapshot-state',
+    '@agent-device/capture-kit/snapshot-state-fixtures',
     '@agent-device/capture-kit/snapshot-timeout-policy',
     '@agent-device/capture-kit/snapshot-visibility',
     '@agent-device/capture-kit/text-surface',
@@ -631,11 +632,12 @@ test('the real tree parses, declares, and passes R11', () => {
   // one subpath per module so consumers pull only the stage they run; the
   // `-fixtures` entries are the test-fixture surface (host-kit's
   // `./audio-probe-fixtures` precedent) — `./interaction-targeting-fixtures`
-  // for the interaction-targeting node trees and `./test-fixtures` for the
-  // snapshot/geometry builders both this package's and root's tests build on
-  // (#2402, replacing the copy that used to live under root's
-  // `src/__tests__/test-utils/`). Any other subpath, or the AST leaking into
-  // `.`, fails here.
+  // for the interaction-targeting node trees and `./snapshot-geometry-fixtures`
+  // for the geometry/touch-point builders both this package's and root's
+  // tests build on (#2402, replacing the copy that used to live under root's
+  // `src/__tests__/test-utils/`; it re-exports `makeSnapshotState` from
+  // capture-kit's own `./snapshot-state-fixtures`, its canonical home). Any
+  // other subpath, or the AST leaking into `.`, fails here.
   assert.deepEqual([...selectorsPackage.exportTargets.keys()].sort(), [
     '@agent-device/selectors',
     '@agent-device/selectors/absence-observation',
@@ -652,7 +654,7 @@ test('the real tree parses, declares, and passes R11', () => {
     '@agent-device/selectors/press-retarget',
     '@agent-device/selectors/selector-pipeline',
     '@agent-device/selectors/selector-pipeline-policy',
-    '@agent-device/selectors/test-fixtures',
+    '@agent-device/selectors/snapshot-geometry-fixtures',
     '@agent-device/selectors/touch-semantics',
   ]);
   assert.deepEqual([...selectorsPackage.workspaceDependencies].sort(), [
