+++ src/commands/interaction/runtime/selector-is.test.ts
@@ -8,7 +8,7 @@ import {
   localCommandPolicy,
 } from '../../../runtime.ts';
 import { selector } from './selector-read-utils.ts';
-import { makeSnapshotState } from '@agent-device/selectors/test-fixtures';
+import { makeSnapshotState } from '@agent-device/selectors/snapshot-geometry-fixtures';
 import { createSelectorDevice, selectorReadSnapshot } from './__tests__/test-utils/index.ts';
 import { AppError } from '@agent-device/kernel/errors';
