+++ src/commands/replay/ad-script-round-trip.test.ts
@@ -2,7 +2,7 @@ import fc from 'fast-check';
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
 import { replayScriptArb } from '../../__tests__/test-utils/property-arbitraries.ts';
-import { PROPERTY_RUNS } from '@agent-device/selectors/test-fixtures';
+import { PROPERTY_RUNS } from '@agent-device/selectors/snapshot-geometry-fixtures';
 import {
   formatPortableActionLine,
   formatTargetAnnotationLines,
