+++ test/integration/interaction-contract/fixtures.ts
@@ -1,5 +1,5 @@
 import type { RawSnapshotNode, SnapshotState } from '@agent-device/kernel/snapshot';
-import { makeSnapshotState } from '@agent-device/selectors/test-fixtures';
+import { makeSnapshotState } from '@agent-device/selectors/snapshot-geometry-fixtures';
 
 /**
  * The permanent contract fixture trees (ADR 0011 Layer 3): the real
